% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_grid.R
\name{abseg}
\alias{abseg}
\title{Add indices as ling segments or points to the figures}
\usage{
abseg(x, y, code = 0, pch = ".", length = 0.05, clear = FALSE, lwd = 1, ...)
}
\arguments{
\item{x}{integer indices, will be used to calculate locations in
\code{\link[graphics]{points}} or \code{\link[graphics]{arrows}}}

\item{y}{y-axis locations for points or line segments}

\item{code, length, ...}{passed to \code{\link[graphics]{arrows}}}

\item{pch}{point styles, passed to \code{\link[graphics]{points}}}

\item{clear}{whether to clear the line before drawing}

\item{lwd}{used to calculate the line weight and point size; see \code{'lwd'}
and \code{'cex'} in \code{\link[graphics]{plot}}, Section 'Details'.}
}
\value{
Nothing.
}
\description{
Draws isolated points as dots, and consecutive points as line
segments.
}
\examples{
plot(1:10)
abseg(c(1,2,3,5,7,8,10), y = 5, lwd = 4, col = 'red')
}
