% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/approx.R
\name{fdp-pwr}
\alias{fdp-pwr}
\alias{fdp}
\alias{pwr}
\title{Calculates false-discovery proportions and statistical power}
\usage{
fdp(rej, support)

pwr(rej, support)
}
\arguments{
\item{rej}{integer indices of rejected hypotheses}

\item{support}{integer indices of the underlying true hypotheses
(with true alternative hypotheses)}
}
\value{
\code{fdp} returns the false-discovery proportions and \code{pwr}
returns statistical power
}
\description{
Calculates false-discovery proportions and statistical power
}
\examples{

# Underlying support is the first 100 hypotheses, but the rejection is
# 2 to 101
support <- 1:100
rejection <- c(2:101)

# Total rejection (length(rejection)) is 100; false rejection is 1
# FDP = 1/100
fdp(rejection, support)

# True positives = 99, total true hypotheses is 100
# power is 99/100
pwr(rejection, support)

}
