% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/focr_grid.R
\name{focr}
\alias{focr}
\alias{focr_initial}
\title{False overlapped-cluster rate (FOCR) control procedures}
\usage{
focr_initial(
  data,
  data_corr,
  blocks,
  nblocks = ncol(data),
  mu = 0,
  alpha = 0.05,
  verbose = FALSE,
  side = c("two", "left", "right"),
  ...
)

focr(
  data,
  block_size,
  alpha = 0.05,
  fdr_method = c("BH", "LAWS", "SABHA", "BY"),
  bandwidth = block_size/2,
  initial_filter = 0.9,
  dimension = NULL,
  distance_measure = c("euclidean", "lmax", "manhattan"),
  side = c("two", "left", "right"),
  verbose = FALSE,
  blocks,
  ...
)
}
\arguments{
\item{data}{a n-by-p numerical matrix (no missing values) with \code{n} to
be the total number of observations and \code{p} is the total number of
hypotheses}

\item{data_corr}{the correlation matrix of \code{data}. If missing, then
the correlation will be calculated empirically}

\item{blocks}{a list of indices or a function that returns indices}

\item{nblocks}{the total number of blocks, used when \code{blocks} is a
function}

\item{mu}{the mean function value to compare with; see 'Details'}

\item{alpha}{FOCR level for stage-I, and FDR level for stage-II}

\item{verbose}{whether to print out information; default is false}

\item{side}{test type, \code{'two'} if alternative hypotheses are two-sided,
and \code{'left'} or \code{'right'} if one-sided.}

\item{...}{passed to \code{focr_initial} and \code{fdr_method}}

\item{block_size}{block size of sliding window; used by \code{focr}.}

\item{fdr_method}{characters or function of post-selection FDR control
procedures. Built-in choices are \code{"BH"}, \code{"BY"}, \code{"SABHA"},
and \code{"LAWS"}. See vignette for details.}

\item{bandwidth}{used by \code{\link{LAWS}} and \code{\link{LAWS}} as
smoothing parameters to estimate the underlying sparsity level.}

\item{initial_filter}{used by \code{\link{LAWS}} and \code{\link{LAWS}} as
initial filters (purity) to remove large p-values}

\item{dimension}{the dimension information of input hypotheses. For
\code{\link{LAWS}} and \code{\link{LAWS}}, current implementation only
supports 1-3 dimensions.}

\item{distance_measure}{distance measure used to form blocks; see 'Details'.}
}
\description{
False overlapped-cluster rate (FOCR) control procedures
}
\details{
The function \code{focr_initial} controls the FOCR on the block level
(stage-I), and calculates the conditional p-values. The function \code{focr}
uses \code{focr_initial}, providing default block settings and built-in
post-selection inference on conditional p-values.

By default, \code{focr} uses sliding window as blocks. Each block is a ball
with distance between the boundary and center point given
by \code{block_size/2}. The distance measure is specified by
\code{distance_measure}. The choices are \code{"euclidean"}, \code{"lmax"},
and \code{"manhattan"}. This default settings should work in many spatial
or temporal situations. However, in case the blocks are to be customized,
please specify \code{blocks} manually. The argument \code{blocks} can be
either a list of hypothesis indices, or a function that returns ones given
by locations of hypotheses. See 'vignette'
}
